% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancelOrder.R
\name{cancelOrder}
\alias{cancelOrder}
\title{Call this function to cancel an order.}
\usage{
cancelOrder(orderId, ib_con)
}
\arguments{
\item{orderId}{The order ID that was specified previously in the \code{\link{placeOrder}}
function}

\item{ib_con}{A valid ib connection.}
}
\description{
Cancels an active order placed by from the same API client ID.
Note: API clients cannot cancel individual orders placed by other clients.
Only reqGlobalCancel can do this.
}
\examples{
Example to follow
}
