% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{Contract}
\alias{Contract}
\title{Create a contract}
\usage{
Contract(conId = 0L, symbol = "", secType = "",
  lastTradeDateOrContractMonth = "", strike = 0, right = "",
  multiplier = "", exchange = "", primaryExchange = "",
  currency = "", localSymbol = "", tradingClass = "",
  includeExpired = FALSE, secIdType = "", secId = "",
  comboLegsDescrip = "", comboLegs = NULL,
  deltaNeutralContract = NULL)
}
\arguments{
\item{conId}{The unique IB contract identifier}

\item{symbol}{The contract (or its underlying) symbol.}

\item{secType}{The security type:}

\item{lastTradeDateOrContractMonth}{The contract's last trading day or
contract month (for Options and Futures).
Strings with format YYYYMM will be interpreted as the Contract Month
whereas YYYYMMDD will be interpreted as Last Trading Day.}

\item{strike}{The option's strike price.}

\item{right}{Put or Call.
Valid values are 'P', 'PUT', 'C', 'CALL', or '' for non-options.}

\item{multiplier}{The instrument's multiplier (i.e. options, futures).}

\item{exchange}{The destination exchange.}

\item{primaryExchange}{The contract's primary exchange. For
smart routed contracts, used to define contract in case of ambiguity.
Should be defined as native exchange of contract, e.g. ISLAND for MSFT.
For exchanges which contain a period in name, will only be part of
exchange name prior to period, i.e. ENEXT for ENEXT.BE.}

\item{currency}{The underlying's currency.}

\item{localSymbol}{The contract's symbol within its primary exchange.
For options, this will be the OCC symbol.}

\item{tradingClass}{The trading class name for this contract.
Available in TWS contract description window as well.}

\item{includeExpired}{If set to true, contract details requests and
historical data queries can be performed pertaining to expired
futures contracts. Expired options or other instrument types are not available.}

\item{secIdType}{Security identifier type}

\item{secId}{Security identifier.}

\item{comboLegsDescrip}{Description of the combo legs.}

\item{comboLegs}{The legs of a combined contract definition.}

\item{deltaNeutralContract}{Delta and underlying price for Delta-Neutral
combo orders.}
}
\value{
returns an object of class "contract"
}
\description{
Create a contract
}
\examples{
ibm <- contract(symbol = "IBM",
                secType = "STK",
                currency = "USD",
                exchange = "SMART")
}
