% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contract.R
\name{ContractDetails}
\alias{ContractDetails}
\title{Object for storing the contract details}
\usage{
ContractDetails(contract = Contract(), marketName = "", minTick = 0,
  orderTypes = "", validExchanges = "", priceMagnifier = 0,
  underConId = 0, longName = "", contractMonth = "", industry = "",
  category = "", subcategory = "", timeZoneId = "",
  tradingHours = "", liquidHours = "", evRule = "",
  evMultiplier = 0, mdSizeMultiplier = 0, aggGroup = 0,
  underSymbol = "", underSecType = "", marketRuleIds = "",
  secIdList = NULL, realExpirationDate = "", lastTradeTime = "",
  cusip = "", ratings = "", descAppend = "", bondType = "",
  couponType = "", callable = FALSE, putable = FALSE, coupon = 0,
  convertible = FALSE, maturity = "", issueDate = "",
  nextOptionDate = "", nextOptionType = "",
  nextOptionPartial = FALSE, notes = "")
}
\arguments{
\item{contract}{The contract information. Detailed information can be found
in the \code{\link{contract}} function}

\item{marketName}{The market name for this product.}

\item{minTick}{The minimum allowed price variation. Note that many securities
vary their minimum tick size according to their price. This value will only
show the smallest of the different minimum tick sizes regardless of the
product's price. Full information about the minimum increment price structure
can be obtained with the reqMarketRule function or the IB Contract
and Security Search site.}

\item{orderTypes}{Supported order types for this product.}

\item{validExchanges}{Valid exchange fields when placing an order for this contract.}

\item{priceMagnifier}{Allows execution and strike prices to be reported
consistently with market data, historical data and the order price,
i.e. Z on LIFFE is reported in Index points and not GBP.}

\item{underConId}{For derivatives, the contract ID (conID) of the underlying instrument.}

\item{longName}{Descriptive name of the product.}

\item{contractMonth}{Typically the contract month of the underlying for a Future contract.}

\item{industry}{The industry classification of the underlying/product. For example, Financial.}

\item{category}{The industry category of the underlying. For example, InvestmentSvc.}

\item{subcategory}{The industry subcategory of the underlying. For example, Brokerage.}

\item{timeZoneId}{The time zone for the trading hours of the product. For example, EST.}

\item{tradingHours}{The trading hours of the product. This value will
contain the trading hours of the current day as well as the next's.}

\item{liquidHours}{The liquid hours of the product. This value will contain
the liquid hours (regular trading hours) of the contract on the specified exchange.}

\item{evRule}{Contains the Economic Value Rule name and the respective optional
argument. The two values should be separated by a colon. For example,
aussieBond:YearsToExpiration=3. When the optional argument is not present,
the first value will be followed by a colon.}

\item{evMultiplier}{Tells you approximately how much the market value of a
contract would change if the price were to change by 1. It cannot be used to
get market value by multiplying the price by the approximate multiplier.}

\item{mdSizeMultiplier}{MD Size Multiplier. Returns the size multiplier for
values returned to tickSize from a market data request.
Generally 100 for US stocks and 1 for other instruments.}

\item{aggGroup}{Aggregated group Indicates the smart-routing group to which
a contract belongs. contracts which cannot be smart-routed have aggGroup = -1.}

\item{underSymbol}{For derivatives, the symbol of the underlying contract.}

\item{underSecType}{For derivatives, returns the underlying security type.}

\item{marketRuleIds}{The list of market rule IDs separated by comma.}

\item{secIdList}{A list of contract identifiers that the customer is allowed to view.}

\item{realExpirationDate}{Real expiration date.}

\item{lastTradeTime}{Last trade time.}

\item{cusip}{The nine-character bond CUSIP. For Bonds only.
Receiving CUSIPs requires a CUSIP market data subscription.}

\item{ratings}{Identifies the credit rating of the issuer.
This field is not currently available from the TWS API.}

\item{descAppend}{A description string containing further descriptive
information about the bond. For Bonds only.}

\item{bondType}{The type of bond, such as "CORP.".}

\item{couponType}{The type of bond coupon. This field is currently not
available from the TWS API. For Bonds only.}

\item{callable}{If true, the bond can be called by the issuer under certain
conditions. This field is currently not available from the TWS API. For Bonds only.}

\item{putable}{Values are True or False. If true, the bond can be sold back
to the issuer under certain conditions. This field is currently not available
 from the TWS API. For Bonds only.}

\item{coupon}{The interest rate used to calculate the amount you will receive
in interest payments over the course of the year.
This field is currently not available from the TWS API. For Bonds only.}

\item{convertible}{Values are True or False. If true, the bond can be
converted to stock under certain conditions. This field is currently not
available from the TWS API. For Bonds only.}

\item{maturity}{The date on which the issuer must repay the face value of
the bond. This field is currently not available from the TWS API. For Bonds only.}

\item{issueDate}{The date the bond was issued. This field is currently not
available from the TWS API. For Bonds only.}

\item{nextOptionDate}{Only if bond has embedded options. This field is
currently not available from the TWS API.}

\item{nextOptionType}{Type of embedded option. This field is currently not
available from the TWS API.}

\item{nextOptionPartial}{Only if bond has embedded options. This field is
currently not available from the TWS API. For Bonds only.}

\item{notes}{If populated for the bond in IB's database. For Bonds only.}
}
\value{
An object of class contract_details
}
\description{
This object is used for storing the contract details when the function
reqContractDetails is used.
}
\examples{
Example to come
}
\keyword{internal}
