% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ib_connect.R
\name{ib_connect}
\alias{ib_connect}
\alias{is_ib_connection}
\alias{ib_disconnect}
\alias{is_ib_connected}
\title{Establish a connection to IB trader workstation or to the IB Gateway}
\usage{
ib_connect(clientId = 1L, host = "127.0.0.1", port = 7497,
  timeout = 5, blocking = .Platform$OS.type == "windows")

is_ib_connection(x)

ib_disconnect(ib_con)

is_ib_connected(ib_con)
}
\arguments{
\item{clientId}{The client ID used in the API connection. Default value is 1.}

\item{host}{Default connection is 127.0.0.1 for the trader workstation or the IB Gateway}

\item{port}{Default connection is to the paper trading account on port 7497.
Live connection to TWS is on port 7496. For the IB gateway the live port is 4001
and the paper trading port is 4002.}

\item{timeout}{Time to establish a connection in seconds. Default is 5.}

\item{blocking}{Create a blocking connection. Defaults to TRUE on Windows, FALSE otherwise.}
}
\value{
Returns an ib_con object.
}
\description{
Establish a connection to IB trader workstation or to the IB Gateway
}
\examples{
\dontrun{
ib_con <- ib_connect()
ib_disconnect(ib_con)
}
}
