% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_type.R
\name{LimitOrder}
\alias{LimitOrder}
\title{Create a limit order object}
\usage{
LimitOrder(action, totalQuantity, lmtPrice)
}
\arguments{
\item{action}{Generally available values are "BUY" and "SELL".
Additionally, "SSHORT", "SLONG" are available in some institutional-accounts only}

\item{totalQuantity}{The number of positions being bought/sold.}

\item{lmtPrice}{The LIMIT price. Used for limit, stop-limit and relative orders}
}
\value{
Returns a limit order object for the placeOrder function.
}
\description{
Create a limit order object
}
\examples{
Example to follow
}
