% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts_by_secType.R
\name{Option}
\alias{Option}
\title{Create an option contract}
\usage{
Option(symbol, lastTradeDateOrContractMonth = "", strike = 0,
  right = "", multiplier = "100", exchange = "SMART",
  primaryExchange = "", currency = "USD")
}
\arguments{
\item{symbol}{The contract (or its underlying) symbol.}

\item{lastTradeDateOrContractMonth}{The contract's last trading day or
contract month.}

\item{strike}{The option's strike price.}

\item{right}{Put or Call. Valid values are 'P', 'PUT', 'C', 'CALL'}

\item{multiplier}{The instrument's multiplier. Default value is: 100}

\item{exchange}{The destination exchange.}

\item{primaryExchange}{The contract's primary exchange.}

\item{currency}{The underlying's currency.}
}
\value{
returns an object of class "Contract"
}
\description{
Create an option contract
}
\examples{
Example to follow
}
\seealso{
Other Contracts: \code{\link{Stock}}
}
\concept{Contracts}
