% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeOrder.R
\name{placeOrder}
\alias{placeOrder}
\title{Place an Order}
\usage{
placeOrder(orderId, contract, order, ib_con)
}
\arguments{
\item{orderId}{The order id. You must specify a unique value. When the
order START_API returns, it will be identified by this tag. This tag is also
used when canceling the order.}

\item{contract}{A valid contract. You can create a valid contract with the
\code{\link{Contract}} function or one of the contract helper functions.}

\item{order}{A valid Order. You can create a valid order with the
\code{\link{Order}} function or one of the order type functions.}

\item{ib_con}{A valid ib connection.}
}
\value{
Returns an order status message
}
\description{
Call this function to place an order. The order status will
be returned by the orderStatus event.
}
\examples{
Example to follow.
}
