% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_messages.R
\name{ProcessMsg}
\alias{ProcessMsg}
\title{Process incoming messages}
\usage{
ProcessMsg(in_msg, ib_con)
}
\arguments{
\item{ib_con}{}
}
\value{
used for processing incoming messages. Called for it's side-effects.
}
\description{
This function is only used for processing incoming messages.
}
\keyword{internal}
