% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqContractDetails.R
\name{reqContractDetails}
\alias{reqContractDetails}
\title{Request contract details}
\usage{
reqContractDetails(contract, reqId = 1L, ib_con)
}
\arguments{
\item{contract}{A valid contract. You can create a valid contract with the
\code{\link{contract}} function or one of the contract helper functions.}

\item{reqId}{The request's unique identifier. Should be a scalar integer.}

\item{ib_con}{A valid ib connection.}
}
\value{
A list of details for the specified contract. Updates nextValidId in
the ib connection.
}
\description{
Request contract details
}
\examples{
\dontrun{
ib_con <- ib_connect()
ibm <- Contract(symbol = "IBM",
                secType = "STK",
                currency = "USD",
                exchange = "SMART")
ibm_details <- reqContractDetails(ibm, reqId = 1L, ib_con = ib_con)
}
}
