% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqCurrentTime.R
\name{reqCurrentTime}
\alias{reqCurrentTime}
\title{Request the current time from the IB server}
\usage{
reqCurrentTime(ib_con)
}
\arguments{
\item{ib_con}{a valid tws connection.}
}
\value{
returns the current time of the IB server
}
\description{
IB's current time. TWS or IB Gateway is synchronized with the server (not local computer)
using NTP and this function will receive the current time in TWS.
}
\examples{
\dontrun{
ib_con <- ib_connect()
reqCurrentTime(ib_con)
}
}
