% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancelOrder.R
\name{reqGlobalCancel}
\alias{reqGlobalCancel}
\title{Call this function to cancel all active orders.}
\usage{
reqGlobalCancel(ib_con)
}
\arguments{
\item{ib_con}{A valid ib connection.}
}
\description{
This method will cancel ALL open orders including those placed directly from TWS.
}
\examples{
Example to follow
}
