% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqIds.R
\name{reqIds}
\alias{reqIds}
\title{Request next valid Id}
\usage{
reqIds(ib_con)
}
\arguments{
\item{ib_con}{a valid ib_con connection object.}
}
\value{
returns the next valid Id.
}
\description{
Requests the next valid order ID at the current moment.
Important: the next valid order ID is only valid at the time it is received.
}
\examples{
\dontrun{
ib_con <- ib_connect()
reqIds(ib_con)
}
}
