% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqMatchingSymbols.R
\name{reqMatchingSymbols}
\alias{reqMatchingSymbols}
\title{Requests matching stock symbols.}
\usage{
reqMatchingSymbols(reqId = 1L, pattern, ib_con)
}
\arguments{
\item{reqId}{id to specify the request. Should be a scalar integer.}

\item{pattern}{Either start of ticker symbol or (for larger strings) company name}

\item{ib_con}{a valid ib_con connection object}
}
\value{
A data.frame containing sample contract descriptions.
}
\description{
Stock Contract Search.
}
\details{
The input can be either the first few letters of the ticker symbol,
or for longer strings, a character sequence matching a word in the security
name. For instance to search for the stock symbol 'IBKR',
the input 'I' or 'IB' can be used, as well as the word 'Interactive'.
Up to 16 matching results are returned.
}
\examples{
\dontrun{
ib_con <- ib_connect()
reqMatchingSymbols(pattern = "Intel", ib_con = ib_con)
}
}
