% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqNewsProviders.R
\name{reqNewsArticle}
\alias{reqNewsArticle}
\title{Request News Article Body}
\usage{
reqNewsArticle(reqId, providerCode, articleId, ib_con)
}
\arguments{
\item{reqId}{id to specify the request. Should be a scalar integer.}

\item{providerCode}{Short code indicating news provider, e.g. FLY}

\item{articleId}{id of the specific article}

\item{ib_con}{A valid ib connection.}
}
\value{
a data.frame containing the article type and article text
}
\description{
Requests news article body given articleId
}
\examples{
Example to follow
}
