% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reqNewsProviders.R
\name{reqNewsProviders}
\alias{reqNewsProviders}
\title{Request news providers}
\usage{
reqNewsProviders(ib_con)
}
\arguments{
\item{ib_con}{A valid ib connection.}
}
\value{
A data.frame with an overview of the news providers
}
\description{
Requests news providers which the user has subscribed to.
}
\examples{
Example to follow
}
