% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts_by_secType.R
\name{Stock}
\alias{Stock}
\title{Create a stock contract}
\usage{
Stock(symbol, exchange = "SMART", primaryExchange = "",
  currency = "USD")
}
\arguments{
\item{symbol}{The contract (or its underlying) symbol.}

\item{exchange}{The destination exchange.}

\item{primaryExchange}{The contract's primary exchange.}

\item{currency}{The underlying's currency.}
}
\value{
returns an object of class "Contract"
}
\description{
Create a stock contract
}
\examples{
Example to follow
}
\seealso{
Other Contracts: \code{\link{Option}}
}
\concept{Contracts}
