% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts_by_secType.R
\name{Option}
\alias{Option}
\title{Create an option contract}
\usage{
Option(symbol, lastTradeDateOrContractMonth = "", strike = 0,
  right = "", multiplier = "100", exchange = "SMART",
  primaryExchange = "", currency = "USD")
}
\arguments{
\item{currency}{}
}
\value{
returns an object of class "contract"
}
\description{
Create an option contract
}
\examples{
Example to follow
}
\seealso{
Other Contracts: \code{\link{Stock}}
}
\concept{Contracts}
