% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts_by_secType.R
\name{Stock}
\alias{Stock}
\title{Create a stock contract}
\usage{
Stock(symbol, exchange = "SMART", primaryExchange = "",
  currency = "USD")
}
\arguments{
\item{currency}{}
}
\value{
returns an object of class "contract"
}
\description{
Create a stock contract
}
\examples{
Example to follow
}
\seealso{
Other Contracts: \code{\link{Option}}
}
\concept{Contracts}
